Sample Text Information
The sample text contains a total of 326 characters. There are 54 different characters in the
text. There are two font runs. The first run contains 115 characters. The second run contains
211 characters.
The following chart indicates the frequency of the characters in the text. For example there is
one character (4) that occurred four times, there are two characters (n, t) that occurred 18
times each in the text.